backend/lambdas/problem-generator-aws/requirements.txt
boto3==1.28.62
langchain==0.0.312
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
